requirements.txt
discord.py==1.6.0
python-dotenv==0.15.0
psycopg2-binary==2.8.6
